Add files via upload
Changes to these files to fix the Heart Beat not working when client's wifi drops or it goes to to sleep causing the server to slow to a crawl.  Heart Beat now works and client disconnects occur quickly with little disruption to other clients.  Tested on ESP8266 with two WS servers and 5 clients.  Clients can now reconnect after wifi drop.  More debug statements added to trace the program flow.